requirements.txt
boto3==1.4.1
botocore==1.4.65
configparser==3.5.0
Django==1.10.2
django-appconf==1.0.2
django-filter==0.15.3
django-pipeline==1.6.9
django-registration-redux==1.4
djangorestframework==3.5.1
docutils==0.12
enum34==1.1.6
flake8==3.0.4
futures==3.0.5
jmespath==0.9.0
mccabe==0.5.2
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
python-dateutil==2.5.3
rcssmin==1.0.6
requests==2.11.1
rjsmin==1.0.12
s3transfer==0.1.9
six==1.10.0
whitenoise==3.2.2
boto3==1.4.1
botocore==1.4.65
configparser==3.5.0
Django==1.10.2
django-appconf==1.0.2
djangorestframework==3.5.1
docutils==0.12
enum34==1.1.6
flake8==3.0.4
futures==3.0.5
jmespath==0.9.0
mccabe==0.5.2
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
python-dateutil==2.5.3
requests==2.11.1
six==1.10.0
whitenoise==3.2.2
